requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
moviepy>=1.0.3
ffmpeg-python>=0.2.0
pydub>=0.25.1
requests>=2.32.0
tqdm>=4.66.0
tenacity>=8.3.0
google-api-python-client>=2.137.0
google-auth>=2.33.0
google-auth-oauthlib>=1.2.0
playwright>=1.47.0
selenium>=4.23.0
pillow>=10.4.0
numpy>=1.26.4
pandas>=2.2.2
